Переименовал таблицу, удалил старые миграции и добавил новую
