--- alembic/versions/2024_08_16_1715-a974f4db4523_test.py
@@ -1,39 +1,0 @@
-"""test
-
-Revision ID: a974f4db4523
-Revises: 
-Create Date: 2024-08-16 17:15:55.446993
-
-"""
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = '9559ab726779'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('user_id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('login', sa.String(length=32), nullable=False),
-    sa.Column('pass_hash', sa.String(length=32), nullable=False),
-    sa.Column('salt', sa.String(), nullable=False),
-    sa.Column('role', sa.Enum('reader', 'writer', 'admin', name='user_roles'), nullable=False),
-    sa.Column('name', sa.String(length=128), nullable=False),
-    sa.PrimaryKeyConstraint('user_id'),
-    sa.UniqueConstraint('login')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('users')
-    # ### end Alembic commands ###
